Do not hide all other layers for ArcGIS service using hiddenids="#ALL#" for FlamingoMapComponent:
ArcGIS server admin should set defaultVisibility property correctly and only for layers which
must be on all the time such as annotation layers.

This change prevents invisible sublayers because a parent layer was set to invisible when 
hiddenids="#ALL#" was being used. We are not gonna set all parent layers to visible to work around
wrong defaultVisibility settings.