make it possible to develop war overlays based on snapshot versions
In the case where a war overlay has a different version number than flamingo the `project.version` of dependencies is set to that of the overlay resulting in build failures. eg. flamingo:4.5.5-SNAPSHOT, overlay: 2.0-SNAPSHOT would (before this change) look for `viewer-config-persistence:2.0-SNAPSHOT`
this adds a property `flamingo.version` to work around that

we should at some point in time probably use the release mechanism to resolve the snapshot version numbers instead