Add event-listening in vector layer for undo/redo/cancel
This adds an event listener when the draw control is activated so drawing can be un-done using `CTRL-Z`, re-done using `CTRL-Y` and cancelled using `ESCAPE` keys. The eventlistener is removed when closing the control.

This affects any control drawing on the OpenlayersVectorLayer (Drawing, Edit, Split) but not he Modify action(as that does not draw). For sanity sake the circle tool is excluded from undo/redo.